requirements.txt
discord.py[voice]
dynaconf
